added an info about not working on windows.

+++ vscode-extension/src/extension.ts
@@ -220,14 +220,22 @@ export function activate(context: vscode.ExtensionContext) {
   let closePort = vscode.commands.registerCommand(
     "extension.closePort",
     async () => {
-      let inputPort = await vscode.window.showInputBox({
-        placeHolder: "Enter the port you need to close?"
-      });
-      let info = await command(`lsof -i :${inputPort}`);
-      let port = transformPort(info);
-      if (port) {
-        await command(`kill -9 ${port}`);
-        vscode.window.setStatusBarMessage("Port closed successfully!");
+      let platform = process.platform;
+
+      if (platform !== "win32") {
+
+        let inputPort = await vscode.window.showInputBox({
+          placeHolder: "Enter the port you need to close?"
+        });
+        let info = await command(`lsof -i :${inputPort}`);
+        let port = transformPort(info);
+        if (port) {
+          await command(`kill -9 ${port}`);
+          vscode.window.setStatusBarMessage("Port closed successfully!");
+        }
+      }
+      else {
+        vscode.window.showErrorMessage('SORRY. Does not work on Windows. But on OSX/Linux.');
       }
     }
   );
